requirements.txt
line-bot-sdk
flask
requests==2.22.0
beautifulsoup4==4.8.2
urllib3  
opencv-python-headless
numpy
gunicorn
